[Feat] finished tags interface

+++ api/src/database/models/post_tags.rs
@@ -7,9 +7,9 @@ use serde::Serialize;
 #[sea_orm(table_name = "post_tags")]
 pub struct Model {
     #[sea_orm(primary_key, auto_increment = false)]
-    pub post_id: i64,
+    pub post_id: i32,
     #[sea_orm(primary_key, auto_increment = false)]
-    pub tag_id: i64,
+    pub tag_id: i32,
 }
 
 #[derive(Copy, Clone, Debug, EnumIter, DeriveRelation)]
